remove input focus zoom and app margins for login
