+++ public/index.html
@@ -4,7 +4,7 @@
 <head>
     <title>Go.gov.sg</title>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <meta http-equiv="Content-Language" content="en">
     <link rel="icon" type="image/png" sizes="16x16" href="/assets/favicon-16x16.png">
